+++ kolobok-annotations/src/main/java/org/kolobok/annotation/DebugLogMask.java
@@ -1,0 +1,13 @@
+package org.kolobok.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.CLASS)
+@Target({ElementType.PARAMETER, ElementType.LOCAL_VARIABLE})
+public @interface DebugLogMask {
+    int first() default 0;
+    int last() default 0;
+}